Create classes folder abd refactor
